reusable

+++ src/main/java/com/naicson/yugioh/data/dto/set/UserSetCollectionDTO.java
@@ -12,7 +12,7 @@ public class UserSetCollectionDTO {
 	private Double totalPrice;
 	private String image;
 	private String setType;
-	private Map<String, Integer> rarities;
+	private Map<String, Long> rarities;
 	private List<String> setCodes;
 	private List<CardSetCollectionDTO> cards;
 	private Map<Long, String> basedDeck;
@@ -35,10 +35,10 @@ public class UserSetCollectionDTO {
 	public void setTotalPrice(Double totalPrice) {
 		this.totalPrice = totalPrice;
 	}
-	public Map<String, Integer> getRarities() {
+	public Map<String, Long> getRarities() {
 		return rarities;
 	}
-	public void setRarities(Map<String, Integer> rarities) {
+	public void setRarities(Map<String, Long> rarities) {
 		this.rarities = rarities;
 	}
 	public List<String> getSetCodes() {
